Starting plots with ticks and ticklabels.

+++ plot/src/svgplot.cc
@@ -2,6 +2,8 @@
 #include "../../src/group.h"
 #include "../../src/rect.h"
 #include "../../src/clipPath.h"
+#include "../../src/line.h"
+#include "../../src/text.h"
 #include <codecvt>
 #include <fstream>
 
@@ -85,6 +87,54 @@ float SVGPlot::linewidth() const noexcept {
     return linewidth_;
 }
 
+std::vector<float> SVGPlot::xticks() const noexcept {
+    if (xticks_) return *xticks_;
+    else if (xticklabels_) { //If there are labels we distribute the ticks uniformly
+        auto [xmin,xmax,d1,d2] = axis();
+        std::vector<float> xt((*xticklabels_).size());
+        float dx = (xmax-xmin)/xt.size();
+        for (std::size_t i = 0;i<xt.size();++i) xt[i] = xmin + (0.5f + float(i))*dx;
+        return xt;
+    } else { //If we don't have any information we rely on the scales
+        auto [xmin,xmax,d1,d2] = axis();
+        return xscale().ticks(5,xmin,xmax);
+    }
+}
+
+std::vector<float> SVGPlot::yticks() const noexcept {
+   if (yticks_) return *yticks_;
+    else if (yticklabels_) { //If there are labels we distribute the ticks uniformly
+        auto [d1,d2,ymin,ymax] = axis();
+        std::vector<float> yt((*yticklabels_).size());
+        float dy = (ymax-ymin)/yt.size();
+        for (std::size_t i = 0;i<yt.size();++i) yt[i] = ymin + (0.5f + float(i))*dy;
+        return yt;
+    } else { //If we don't have any information we rely on the scales
+        auto [d1,d2,ymin,ymax] = axis();
+        return yscale().ticks(5,ymin,ymax);
+    }
+}
+
+std::vector<std::string> SVGPlot::xticklabels() const noexcept {
+    if (xticklabels_) return *xticklabels_;
+    else {
+        auto xt = xticks();
+        std::vector<std::string> xtl(xt.size());
+        for (std::size_t i = 0; i<xtl.size(); ++i) xtl[i] = xscale().ticklabel(xt[i]);
+        return xtl;
+    }
+}
+
+std::vector<std::string> SVGPlot::yticklabels() const noexcept {
+    if (yticklabels_) return *yticklabels_;
+    else {
+        auto yt = yticks();
+        std::vector<std::string> ytl(yt.size());
+        for (std::size_t i = 0; i<ytl.size(); ++i) ytl[i] = yscale().ticklabel(yt[i]);
+        return ytl;
+    }
+}
+
 svg::Color SVGPlot::next_color() noexcept {
     auto c = color_cycle[color_cycle_pos];
     color_cycle_pos = (color_cycle_pos + 1)%color_cycle.size();
@@ -92,11 +142,16 @@ svg::Color SVGPlot::next_color() noexcept {
 }
 
 std::array<float,4> SVGPlot::margin() const noexcept {
+    bool xt = !xticks().empty(); bool xtl = !xticklabels().empty();
+    bool yt = !yticks().empty(); bool ytl = !yticklabels().empty();
+    //Assuming font size of 10 but if we add fonts this needs to change
+
+    //Right and top are affected by xticklabels and yticklabels because the text might get out of there
     return std::array<float,4>{
-        0.5f*linewidth(), //Left
-        0.5f*linewidth(), //Right
-        0.5f*linewidth(), //Top
-        0.5f*linewidth()  //Bottom
+        0.5f*linewidth() + (xt?3.0f:0.0f) + (xtl?(3.0f*10.0f/2.0f):0.0f), //Left
+        0.5f*linewidth() + (ytl?10.0f:0.0f), //Right
+        0.5f*linewidth() + (xtl?7.5f:0.0f), //Top
+        0.5f*linewidth() + (yt?3.0f:0.0f) + (ytl?15.0f:0.0f) //Bottom
     };
 }
 
@@ -111,6 +166,31 @@ SVGPlot& SVGPlot::axis(const std::array<float,4>& a) noexcept {
 SVGPlot& SVGPlot::linewidth(float lw) noexcept {
     linewidth_ = lw; return (*this);
 }
+
+SVGPlot& SVGPlot::xticks(const std::vector<float>& xt) noexcept {
+    xticks_ = xt; return (*this);
+}
+
+SVGPlot& SVGPlot::yticks(const std::vector<float>& yt) noexcept {
+    yticks_ = yt; return (*this);
+}
+
+SVGPlot& SVGPlot::xticklabels(const std::vector<std::string>& xtl) noexcept {
+    xticklabels_ = xtl; return (*this);
+}
+
+SVGPlot& SVGPlot::yticklabels(const std::vector<std::string>& ytl) noexcept {
+    yticklabels_ = ytl; return (*this);
+}
+
+SVGPlot& SVGPlot::xticks(const std::vector<float>& xt, const std::vector<std::string>& xtl) noexcept {
+    return xticks(xt).xticklabels(xtl);
+}
+
+SVGPlot& SVGPlot::yticks(const std::vector<float>& yt, const std::vector<std::string>& ytl) noexcept {
+    return yticks(yt).yticklabels(ytl);
+}
+
 
 
     /****************
@@ -129,9 +209,32 @@ svg::SVG SVGPlot::svg() const noexcept {
     svg::Group& plots = output.add(svg::Group());
     plots.clip_path(svg::Url(clip.id()));
     for (const Plottable& plottable : plottables)
-        plots.add(plottable.plot(xscale().transform(ax[0],ax[1],mg[0],fs[0]-mg[1]),
-                                 yscale().transform(ax[2],ax[3],mg[2],fs[1]-mg[3])));
+        plots.add(plottable.plot(xscale().transform(ax[0],ax[1],border.x(),border.x()+border.width()),
+                                 yscale().transform(ax[2],ax[3],border.y()+border.height(),border.y())));
     output.add(border).fill(svg::none).stroke(svg::black).stroke_width(linewidth());
+
+    auto xt = xticks();
+    auto xtl = xticklabels();
+    //We don't give the option of configuring tick appearance yet. Width = linewidth and length = 3
+    const float ticklength = 3.0f;
+    for (std::size_t i = 0; i<xt.size(); ++i) {
+        float x = xscale().transform(xt[i],ax[0],ax[1],border.x(),border.x()+border.width());
+        output.add(svg::Line(x,border.y()+border.height(),x,border.y()+border.height()+ticklength)).stroke_width(linewidth()).stroke(svg::black);
+        if (i<xtl.size())
+            output.add(svg::Text(xtl[i],x,border.y()+border.height()+ticklength+1.0f))
+                .font_size(10.0f).font_family("sans-serif").dominant_baseline(svg::hanging).text_anchor(svg::middle);
+    }
+
+    auto yt = yticks();
+    auto ytl = yticklabels();
+    for (std::size_t i = 0; i<yt.size(); ++i) {
+        float y = yscale().transform(yt[i],ax[2],ax[3],border.y()+border.height(),border.y());
+        output.add(svg::Line(border.x(),y,border.x()-ticklength,y)).stroke_width(linewidth()).stroke(svg::black);
+        if (i<ytl.size())
+            output.add(svg::Text(ytl[i],border.x()-ticklength-1.0f,y))
+                .font_size(10.0f).font_family("sans-serif").dominant_baseline(svg::middle).text_anchor(svg::end);
+    }
+
     return output;
 }
 
@@ -148,16 +251,18 @@ void SVGPlot::savefig(const std::filesystem::path& name) const {
     std::ofstream f(svg_name);
     svg().save_to_stream(f);
     f.close();
+    int status = -1;
     if (name.extension() == ".png") {
-        std::system((std::string("inkscape --export-type=\"png\" ")+convert(svg_name.native())).c_str());
-        std::filesystem::remove(svg_name);
+        status = std::system((std::string("inkscape --export-type=\"png\" ")+convert(svg_name.native())).c_str());
     } else if (name.extension() == ".pdf") {
-        std::system((std::string("inkscape --export-type=\"pdf\" ")+convert(svg_name.native())).c_str());
-        std::filesystem::remove(svg_name);
+        status = std::system((std::string("inkscape --export-type=\"pdf\" ")+convert(svg_name.native())).c_str());
     } else if (name.extension() == ".eps") {
-        std::system((std::string("inkscape --export-type=\"eps\" ")+convert(svg_name.native())).c_str());
-        std::filesystem::remove(svg_name);
+        status = std::system((std::string("inkscape --export-type=\"eps\" ")+convert(svg_name.native())).c_str());
+    } else if (name.extension() == ".pdf_tex") {
+        status = std::system((std::string("inkscape --export-type=\"pdf\" --export-latex ")+convert(svg_name.native())).c_str());
     }
+
+    if (status == 0) std::filesystem::remove(svg_name);
 }	
 
 }
